requirements.txt
streamlit>=1.25.0
mysql-connector-python>=8.0.0
pandas>=1.4.0
python-dotenv>=0.20.0
